requirements-travis.txt
pytest==3.0.2
Sphinx==1.4.6
PyYAML==3.12
ipython
